ApplicationDir moved to fpath
