+++ pkg/process/exec_conf.go
@@ -9,8 +9,6 @@ import (
 	"fmt"
 	"os"
 	"os/signal"
-	"path/filepath"
-	"runtime"
 	"strings"
 	"sync"
 	"syscall"
@@ -196,41 +194,4 @@ func cleanup(cmd *cobra.Command) {
 		}
 		return err
 	}
-}
-
-// ApplicationDir returns best base directory for specific OS
-func ApplicationDir(subdir ...string) string {
-	for i := range subdir {
-		if runtime.GOOS == "windows" || runtime.GOOS == "darwin" {
-			subdir[i] = strings.Title(subdir[i])
-		} else {
-			subdir[i] = strings.ToLower(subdir[i])
-		}
-	}
-	var appdir string
-	home := os.Getenv("HOME")
-
-	switch runtime.GOOS {
-	case "windows":
-		// Windows standards: https://msdn.microsoft.com/en-us/library/windows/apps/hh465094.aspx?f=255&MSPPError=-2147217396
-		for _, env := range []string{"AppData", "AppDataLocal", "UserProfile", "Home"} {
-			val := os.Getenv(env)
-			if val != "" {
-				appdir = val
-				break
-			}
-		}
-	case "darwin":
-		// Mac standards: https://developer.apple.com/library/archive/documentation/FileManagement/Conceptual/FileSystemProgrammingGuide/MacOSXDirectories/MacOSXDirectories.html
-		appdir = filepath.Join(home, "Library", "Application Support")
-	case "linux":
-		fallthrough
-	default:
-		// Linux standards: https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html
-		appdir = os.Getenv("XDG_DATA_HOME")
-		if appdir == "" && home != "" {
-			appdir = filepath.Join(home, ".local", "share")
-		}
-	}
-	return filepath.Join(append([]string{appdir}, subdir...)...)
 }
